fix: update exception_handler
